+++ internal/helper/housekeeping/housekeeping.go
@@ -20,20 +20,19 @@ func Perform(ctx context.Context, repoPath string) error {
 	filesMarkedForRemoval := 0
 	unremovableFiles := 0
 
-	err := filepath.Walk(repoPath, func(path string, info os.FileInfo, err error) error {
+	err := filepath.Walk(repoPath, func(path string, info os.FileInfo, _ error) error {
 		if repoPath == path {
 			// Never consider the root path
 			return nil
 		}
 
-		if info == nil || !shouldRemove(path, info.ModTime(), info.Mode(), err) {
+		if info == nil || !shouldRemove(path, info.ModTime(), info.Mode()) {
 			return nil
 		}
 
 		filesMarkedForRemoval++
 
-		err = forceRemove(path)
-		if err != nil {
+		if err := forceRemove(path); err != nil {
 			unremovableFiles++
 			logger.WithError(err).WithField("path", path).Warn("unable to remove stray file")
 		}
@@ -56,10 +55,29 @@ func Perform(ctx context.Context, repoPath string) error {
 	return err
 }
 
-func fixPermissions(path string) {
-	filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
-		if info.IsDir() {
-			os.Chmod(path, 0700)
+// FixDirectoryPermissions does a recursive directory walk to look for
+// directories that cannot be accessed by the current user, and tries to
+// fix those with chmod. The motivating problem is that directories with mode
+// 0 break os.RemoveAll.
+func FixDirectoryPermissions(path string) error {
+	return fixDirectoryPermissions(path, make(map[string]struct{}))
+}
+
+const minimumDirPerm = 0700
+
+func fixDirectoryPermissions(path string, retriedPaths map[string]struct{}) error {
+	return filepath.Walk(path, func(path string, info os.FileInfo, errIncoming error) error {
+		if !info.IsDir() || info.Mode()&minimumDirPerm == minimumDirPerm {
+			return nil
+		}
+
+		if err := os.Chmod(path, info.Mode()|minimumDirPerm); err != nil {
+			return err
+		}
+
+		if _, retried := retriedPaths[path]; !retried && os.IsPermission(errIncoming) {
+			retriedPaths[path] = struct{}{}
+			return fixDirectoryPermissions(path, retriedPaths)
 		}
 
 		return nil
@@ -74,12 +92,14 @@ func forceRemove(path string) error {
 	}
 
 	// Delete failed. Try again after chmod'ing directories recursively
-	fixPermissions(path)
+	if err := FixDirectoryPermissions(path); err != nil {
+		return err
+	}
 
 	return os.RemoveAll(path)
 }
 
-func shouldRemove(path string, modTime time.Time, mode os.FileMode, err error) bool {
+func shouldRemove(path string, modTime time.Time, mode os.FileMode) bool {
 	base := filepath.Base(path)
 
 	// Only delete entries starting with `tmp_` and older than a week