+++ internal/service/repository/calculate_checksum.go
@@ -1,10 +1,10 @@
 package repository
 
-// Stubbed out until https://gitlab.com/gitlab-org/gitaly/merge_requests/642 is done
-
 import (
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+
 	"gitlab.com/gitlab-org/gitaly/internal/helper"
+
 	"golang.org/x/net/context"
 )