+++ internal/tempdir/tempdir.go
@@ -2,17 +2,16 @@ package tempdir
 
 import (
 	"context"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path"
-	"path/filepath"
 	"strings"
 	"time"
 
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
 	"gitlab.com/gitlab-org/gitaly/internal/config"
 	"gitlab.com/gitlab-org/gitaly/internal/helper"
+	"gitlab.com/gitlab-org/gitaly/internal/helper/housekeeping"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -22,6 +21,8 @@ const (
 	// directory name that could be provided by a user. The '+' character is
 	// not allowed in GitLab namespaces or repositories.
 	tmpRootPrefix = "+gitaly/tmp"
+
+	maxAge = 7 * 24 * time.Hour
 )
 
 // New returns the path of a new temporary directory for use with the
@@ -76,6 +77,8 @@ func StartCleaning() {
 	}
 }
 
+type invalidCleanRoot string
+
 func clean(dir string) error {
 	// If we start "cleaning up" the wrong directory we may delete user data
 	// which is Really Bad.
@@ -84,37 +87,27 @@ func clean(dir string) error {
 		panic(invalidCleanRoot("invalid tempdir clean root: panicking to prevent data loss"))
 	}
 
-	return filepath.Walk(dir, cleanFunc)
-}
-
-const (
-	maxAge = 7 * 24 * time.Hour
-)
-
-type invalidCleanRoot string
-
-func cleanFunc(path string, info os.FileInfo, errIncoming error) error {
-	if errIncoming != nil && !os.IsNotExist(errIncoming) {
-		return fmt.Errorf("incoming %q: %v", path, errIncoming)
-	}
-
-	if info == nil {
+	entries, err := ioutil.ReadDir(dir)
+	if os.IsNotExist(err) {
 		return nil
 	}
+	if err != nil {
+		return err
+	}
 
-	if perm := info.Mode().Perm(); info.IsDir() && perm&0700 < 0700 {
-		// Fix directory read permissions
-		if err := os.Chmod(path, perm|0700); err != nil {
+	for _, info := range entries {
+		if time.Since(info.ModTime()) < maxAge {
+			continue
+		}
+
+		fullPath := path.Join(dir, info.Name())
+		if err := housekeeping.FixDirectoryPermissions(fullPath); err != nil {
 			return err
 		}
-	}
 
-	if time.Since(info.ModTime()) < maxAge {
-		return nil
-	}
-
-	if err := os.Remove(path); err != nil && !info.IsDir() {
-		return err
+		if err := os.RemoveAll(fullPath); err != nil {
+			return err
+		}
 	}
 
 	return nil