+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 UNRELEASED
 
+- Implement Get{Tag,Commit}Messages RPCs
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/646
+- Fix directory permission walker for Go 1.10
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/650
+
+v0.93.0
+
 - Fix concurrency limit handler stream interceptor
   https://gitlab.com/gitlab-org/gitaly/merge_requests/640
 - Vendor gitlab_git at 9b76d8512a5491202e5a953